decorator: Extract zerver/lib/profile.py.
This lets us remove this debugging function from our core
authentication codebase.
